requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.3
pydantic-settings==2.9.1
python-dotenv==1.1.0
orjson==3.10.16
redis==6.0.0
psycopg[binary]==3.2.6
cassandra-driver==3.29.2
kafka-python==2.1.4
pandas==2.2.3
numpy==2.2.5
yfinance==0.2.57
httpx==0.28.1
pytest==8.3.5

